requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
openai>=1.0,<2.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0

